# src/main/frontend/public/openapi.yaml
...
openapi: 3.0.3
info:
  title: Hamstrack REST API
  version: 0.2.0
  description: |
    The HTTP API behind the Hamstrack web app — everything the UI does goes
    through it, and it is available to users for scripting and integrations.

    **Authentication** — call `POST /api/auth/login` and send the returned
    access token as `Authorization: Bearer <token>` (use the **Authorize**
    button here). Tokens expire after 24 h; refresh via the rotating
    `refresh_token` cookie with `POST /api/auth/refresh`.

    **Access model** — a resource you cannot see returns `404`, whether it
    does not exist or you are not a member of its workspace. Lists are
    currently unpaginated (beta). The API is unversioned while in beta;
    breaking changes are announced in release notes.

    **Rate limits** — sensitive auth endpoints share a per-IP budget
    (15 requests/min by default), and repeated failed logins for one account
    trigger an exponential backoff. Both return `429` with a `Retry-After`
    header (seconds).
  contact:
    name: Hamstrack
    url: https://github.com/Zherikhov/easyTask
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: https://hamstrack.com/api
    description: Cloud
  - url: /api
    description: This instance (Cloud or self-hosted DC)
...
security:
  - bearerAuth: []
paths:
...
  /workspaces/{workspaceId}/projects/{projectId}/issues:
    post:
      tags: [Issues]
      summary: Create an issue
      description: |
        `assigneeId` must be a workspace member; `parentId` links a sub-task
        to a parent issue in the same project. The issue gets the next
        sequential project-scoped `number`.
      operationId: createIssue
      parameters:
        - $ref: '#/components/parameters/workspaceId'
        - $ref: '#/components/parameters/projectId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateIssueRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IssueResponse' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Project is archived
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetail' }
        '422': { $ref: '#/components/responses/UnprocessableReference' }
    get:
      tags: [Issues]
      summary: List issues
      operationId: listIssues
      parameters:
        - $ref: '#/components/parameters/workspaceId'
        - $ref: '#/components/parameters/projectId'
        - name: statusId
          in: query
          schema: { type: string, format: uuid }
        - name: assigneeId
          in: query
          schema: { type: string, format: uuid }
        - name: priorityId
          in: query
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/IssueResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
  /workspaces/{workspaceId}/projects/{projectId}/issues/{number}:
    get:
      tags: [Issues]
      summary: Get an issue
      operationId: getIssue
      parameters:
        - $ref: '#/components/parameters/workspaceId'
        - $ref: '#/components/parameters/projectId'
        - $ref: '#/components/parameters/issueNumber'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IssueResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Issues]
      summary: Update an issue
      description: |
        Partial update — send any subset of fields plus the `version` you last
        read; a stale version returns `409` (omit `version` to skip the check).
        Status changes are validated against workflow transition rules (`422`
        when forbidden).
      operationId: updateIssue
      parameters:
        - $ref: '#/components/parameters/workspaceId'
        - $ref: '#/components/parameters/projectId'
        - $ref: '#/components/parameters/issueNumber'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateIssueRequest' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IssueResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Stale `version` — the issue was modified by someone else, or the project is archived
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetail' }
        '422': { $ref: '#/components/responses/UnprocessableReference' }
    delete:
      tags: [Issues]
      summary: Delete an issue
      description: Requires project `MANAGER`. Removes comments and attachment files too.
      operationId: deleteIssue
      parameters:
        - $ref: '#/components/parameters/workspaceId'
        - $ref: '#/components/parameters/projectId'
        - $ref: '#/components/parameters/issueNumber'
      responses:
        '204': { description: Deleted }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /workspaces/{workspaceId}/projects/{projectId}/issues/{number}/history:
    get:
      tags: [Issues]
      summary: Field-level change history
      operationId: getIssueHistory
      parameters:
        - $ref: '#/components/parameters/workspaceId'
        - $ref: '#/components/parameters/projectId'
        - $ref: '#/components/parameters/issueNumber'
      responses:
        '200':
          description: OK, oldest first
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/IssueHistoryResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
  /workspaces/{workspaceId}/projects/{projectId}/issues/{number}/comments:
    post:
      tags: [Comments]
      summary: Add a comment
      description: '`@DisplayName` mentions notify the mentioned workspace members.'
      operationId: createComment
      parameters:
        - $ref: '#/components/parameters/workspaceId'
        - $ref: '#/components/parameters/projectId'
        - $ref: '#/components/parameters/issueNumber'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateCommentRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CommentResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
    get:
      tags: [Comments]
      summary: List comments
      description: Soft-deleted comments are excluded.
      operationId: listComments
      parameters:
        - $ref: '#/components/parameters/workspaceId'
        - $ref: '#/components/parameters/projectId'
        - $ref: '#/components/parameters/issueNumber'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/CommentResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
  /workspaces/{workspaceId}/projects/{projectId}/issues/{number}/comments/{commentId}:
    patch:
      tags: [Comments]
      summary: Edit a comment
      description: Author only.
      operationId: updateComment
      parameters:
        - $ref: '#/components/parameters/workspaceId'
        - $ref: '#/components/parameters/projectId'
        - $ref: '#/components/parameters/issueNumber'
        - $ref: '#/components/parameters/commentId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateCommentRequest' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CommentResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Comments]
      summary: Delete a comment (soft)
      description: Author only.
      operationId: deleteComment
      parameters:
        - $ref: '#/components/parameters/workspaceId'
        - $ref: '#/components/parameters/projectId'
        - $ref: '#/components/parameters/issueNumber'
        - $ref: '#/components/parameters/commentId'
      responses:
        '204': { description: Deleted }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /workspaces/{workspaceId}/projects/{projectId}/issues/{number}/attachments:
    post:
      tags: [Attachments]
      summary: Upload an attachment
      description: '`multipart/form-data` with a `file` field. Default size limit 25 MB (`413` when exceeded).'
      operationId: uploadAttachment
      parameters:
        - $ref: '#/components/parameters/workspaceId'
        - $ref: '#/components/parameters/projectId'
        - $ref: '#/components/parameters/issueNumber'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Uploaded
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AttachmentResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
        '413':
          description: File exceeds the upload size limit
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetail' }
    get:
      tags: [Attachments]
      summary: List attachments
      operationId: listAttachments
      parameters:
        - $ref: '#/components/parameters/workspaceId'
        - $ref: '#/components/parameters/projectId'
        - $ref: '#/components/parameters/issueNumber'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AttachmentResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
  /workspaces/{workspaceId}/projects/{projectId}/issues/{number}/attachments/{attachmentId}:
    get:
      tags: [Attachments]
      summary: Download an attachment
      operationId: downloadAttachment
      parameters:
        - $ref: '#/components/parameters/workspaceId'
        - $ref: '#/components/parameters/projectId'
        - $ref: '#/components/parameters/issueNumber'
        - $ref: '#/components/parameters/attachmentId'
      responses:
        '200':
          description: 'Binary content, `Content-Disposition: attachment`'
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Attachments]
      summary: Delete an attachment
      description: Uploader or project `MANAGER`.
      operationId: deleteAttachment
      parameters:
        - $ref: '#/components/parameters/workspaceId'
        - $ref: '#/components/parameters/projectId'
        - $ref: '#/components/parameters/issueNumber'
        - $ref: '#/components/parameters/attachmentId'
      responses:
        '204': { description: Deleted }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token from `POST /api/auth/login` (24 h lifetime).
    cookieAuth:
      type: apiKey
      in: cookie
      name: refresh_token
      description: Rotating refresh token cookie set on login (30-day lifetime, path `/api/auth`).
  parameters:
    workspaceId:
      name: workspaceId
      in: path
      required: true
      schema: { type: string, format: uuid }
    projectId:
      name: projectId
      in: path
      required: true
      schema: { type: string, format: uuid }
    issueNumber:
      name: number
      in: path
      required: true
      description: Project-scoped issue number (the `42` in `DEMO-42`)
      schema: { type: integer, format: int64 }
    commentId:
      name: commentId
      in: path
      required: true
      schema: { type: string, format: uuid }
    attachmentId:
      name: attachmentId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    Forbidden:
      description: Authenticated but not allowed (insufficient role)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
    NotFound:
      description: Not found — or not a member of the containing workspace
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
    ValidationError:
      description: Failed validation — see the `errors` map
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ValidationProblemDetail' }
    UnprocessableReference:
      description: Unknown status/type/assignee/parent, or a workflow-forbidden transition
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
...
  schemas:
    ProblemDetail:
      type: object
      description: RFC 9457 problem details
      properties:
        type: { type: string, example: 'about:blank' }
        title: { type: string, example: Conflict }
        status: { type: integer, example: 409 }
        detail: { type: string, example: Issue was modified by someone else — refresh and retry }
    ValidationProblemDetail:
      allOf:
        - $ref: '#/components/schemas/ProblemDetail'
        - type: object
          properties:
            errors:
              type: object
              additionalProperties: { type: string }
              example: { email: must be a well-formed email address }
...
    Priority:
      type: object
      description: Catalog entity (was a closed enum before M1)
      properties:
        id: { type: string, format: uuid }
        name: { type: string, example: High }
        color: { type: string, example: '#EA580C' }
        icon: { type: string, nullable: true, example: chevron-up, description: lucide icon name }
        position: { type: integer }
    StatusCategory:
      type: string
      enum: [TODO, IN_PROGRESS, DONE]
      description: Drives board grouping and backlog filtering

    # ---- Admin console schemas ----
...
    FieldValueEntry:
      type: object
      description: A filled custom field on an issue
      properties:
        fieldId: { type: string, format: uuid }
        value: { description: Shape depends on the field type — see FieldType }
...
    IssueTypeResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string, example: Bug }
        color: { type: string, example: '#EF4444' }
        icon: { type: string, nullable: true, example: bug }
        position: { type: integer }
    StatusResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string, example: In Progress }
        color: { type: string, example: '#3B82F6' }
        category: { $ref: '#/components/schemas/StatusCategory' }
        position: { type: integer }
...
    UserRef:
      type: object
      nullable: true
      properties:
        id: { type: string, format: uuid }
        displayName: { type: string }
        avatarUrl: { type: string, nullable: true }
    CreateIssueRequest:
      type: object
      required: [title, typeId, statusId]
      properties:
        title: { type: string, maxLength: 500 }
        description: { type: string, nullable: true }
        typeId: { type: string, format: uuid, description: Must be offered by the project's type set }
        statusId: { type: string, format: uuid, description: Must belong to the project's workflow }
        priorityId: { type: string, format: uuid, nullable: true, description: Omitted = the project's default priority }
        assigneeId: { type: string, format: uuid, nullable: true, description: Must be a workspace member }
        parentId: { type: string, format: uuid, nullable: true, description: Parent issue in the same project }
        dueDate: { type: string, format: date, nullable: true }
        fields:
          type: object
          nullable: true
          additionalProperties: true
          description: >-
            Custom field values keyed by field id (value shapes per
            FieldType). Required fields of the project's set must be
            present; fields outside the set or archived are rejected (422).
          example: { 'e1b2…': 5, 'f3c4…': critical }
    UpdateIssueRequest:
      type: object
      description: Partial update — omitted fields are left unchanged
      properties:
        title: { type: string, minLength: 1, maxLength: 500 }
        description: { type: string }
        typeId: { type: string, format: uuid, description: Must be offered by the project's type set }
        statusId: { type: string, format: uuid, description: Validated against workflow transition rules }
        priorityId: { type: string, format: uuid, description: Must be offered by the project's priority set }
        assigneeId: { type: string, format: uuid }
        dueDate: { type: string, format: date }
        fields:
          type: object
          additionalProperties: true
          description: >-
            Partial like the rest of the body — only listed field ids
            change; JSON null clears a value (required fields cannot be
            cleared).
        version:
          type: integer
          description: Optimistic lock — the `version` you last read; `409` when stale. Omit to skip the check.
    IssueResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        number: { type: integer, format: int64, example: 18 }
        key: { type: string, example: DEMO-18 }
        title: { type: string }
        description: { type: string, nullable: true }
        type: { $ref: '#/components/schemas/IssueTypeResponse' }
        status: { $ref: '#/components/schemas/StatusResponse' }
        priority: { $ref: '#/components/schemas/Priority' }
        assignee: { $ref: '#/components/schemas/UserRef' }
        reporter: { $ref: '#/components/schemas/UserRef' }
        parentId: { type: string, format: uuid, nullable: true }
        dueDate: { type: string, format: date, nullable: true }
        fields:
          type: array
          description: Filled custom fields only
          items: { $ref: '#/components/schemas/FieldValueEntry' }
        version: { type: integer, description: Send back in PATCH for optimistic locking }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    IssueHistoryResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        field: { type: string, example: status, description: 'Built-in attribute name, or the display name of a custom field' }
        oldValue: { type: string, nullable: true, example: To Do }
        newValue: { type: string, nullable: true, example: In Progress }
        changedById: { type: string, format: uuid }
        changedByName: { type: string }
        createdAt: { type: string, format: date-time }
    CreateCommentRequest:
      type: object
      required: [body]
      properties:
        body: { type: string, description: '`@DisplayName` mentions notify workspace members' }
    CommentResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        authorId: { type: string, format: uuid }
        authorName: { type: string }
        body: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    AttachmentResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        filename: { type: string, example: screenshot.png }
        sizeBytes: { type: integer, format: int64 }
        contentType: { type: string, example: image/png }
        uploadedById: { type: string, format: uuid }
        uploadedByName: { type: string }
        createdAt: { type: string, format: date-time }
...